feat: edit and delete existing Series

+++ src/main/java/com/tuka/comiccharacters/dao/SeriesDaoImpl.java
@@ -10,7 +10,7 @@ public class SeriesDaoImpl extends AbstractJpaDao<Series> {
         super(Series.class);
     }
 
-    public Series findByIdWithIssues(int id) {
+    public Series findByIdWithIssues(Long id) {
         EntityManager em = getEntityManager();
         try {
             return em.createQuery(
